requirements.txt
smolagents
langchain
google-generativeai
